backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2==2.9.9
bcrypt==4.1.2
pytest-django==4.8.0
pytest==8.2.0
pytest-cov==5.0.0
factory_boy==3.3.0
djangorestframework-simplejwt==5.3.1
pytz==2024.1